fix type issues when compiling with msvc

+++ src/samples/raytracing/simple_render_rt.cpp
@@ -87,10 +87,11 @@ void SimpleRender::RayTraceCPU()
   }
 
   m_pRayTracerCPU->UpdateView(m_cam.pos, m_inverseProjViewMatrix);
-#pragma omp parallel for default(none)
-  for (size_t j = 0; j < m_height; ++j)
+
+  #pragma omp parallel for default(none)
+  for (int j = 0; j < m_height; ++j)
   {
-    for (size_t i = 0; i < m_width; ++i)
+    for (int i = 0; i < m_width; ++i)
     {
       m_pRayTracerCPU->CastSingleRay(i, j, m_raytracedImageData.data());
     }
